domain_multi : remove superfluous per-proc writing of rank_?/nr_patches
This was only being written/read if the mrc_io type wasn't parallel, but
since nr_patches is a param for domain_multi it was already being written,
meaning this information was stored without the "rank" bit. In serial io
there isn't a chance of this value getting clobbered, so there was no
reason to write the per patch ranks in the first place.

Relevance: adios can pretend it's a serial io method (even though it's
parallel) by choosing to read only information from a certain block.
The issue is (and this is kind of a bad thing in adios) the blocks start counting
at 0 for each processor that writes the variable. If every proc writes it, then
the writeblocks line up with the rank. If, however, only proc 2 writes a variable
then the writeblock of proc 2 for that variable will be 0, and a read using writeblock
2 will give 0.

Long term, a better solution to the "if (io->parallel) then name="rank_%d" check
would be to institute a mrc_io_write_attr_safe that is guaranteed not to clobber
no matter what io system is used. see ticket:26