io -> adios : change the read in selection scope based on the number of writeblocks
PSC has some local objects (ie, patches for the fields and particles) which need to
be read with selection NULL instead of the writeblock selection. So I added in a check
that will set selection to NULL if there is only one writeblock.

For safety there's also an assert that the number of writeblocks (if > 1) is the same
as the size of the io objects.