checkpoint : hack around adios' mrc_fld related segfault on chester
I don't know if this is a problem with adios, the build on chester,
or my code, but writing mrc_fld's with 5D offset/dim/gdim lists
seems to cause a segfault deep within adios, even in a small breakout
test problem. It looks like doing a straight dump/read of data of length _len
works, though.