io : aidos : Handle cases where the same variable is written from a subset of processors
For PSC checkpointing a bunch of local mrc_objs are written to disk with uids based on their
physical memory address. For even small runs these uids may not actually be unique, resulting
in the same variable being written from a subset of processors. We check this, and associate the
correct writeblock with each reader.

WARNING: this will only work with ADIOS > 1.9.0. At the time of this commit the needed
features are only implemented in the master branch of the adios git repo. Unfortunately
that version of the code still reports the adios_version as 1.9.0, so we can't update configure.ac
to block older versions. When the next release of ADIOS is given a number we'll need to set
that as the minimum version.